Refactor string_tools; move createCStyleArgs function to string_tools for better modularity and reuse
