+++ include/tools/string_tools.h
@@ -6,7 +6,8 @@
 namespace tools::string
 {
     std::vector<std::string> splitString(const std::string &input, char delimiter = ' ');
-    pid_t                    strToPid(const std::string &pidStr);   
+    pid_t                    strToPid(const std::string &pidStr);
+    std::vector<char*> createCStyleArgs(const std::vector<std::string>& args);
 }
 
 #endif // STRING_TOOLS_H
